Remove ACID aside from workbench
I'm not sure if this will be controversial,
but I'd like to see this go away...

While it's cool to demo the Asides feature, I've always found that it
distracts from demoing the actual XBlocks.
- It can confuse new developers who wonder where the functionality is
  coming from.
- It can confuse our product owners when demoing XBlocks, as we use the
  workbench to demo new functionality.

I'm open to other ways of documenting/demonstrating this behavior,
but I'd like to remove it from the default case.