Run make upgrade.
Update quality to build off of test.txt since import checks look to see
if various libraries are there.  Using test.txt instead of test.in
because we want qualty to pin at the same version as test.

Reorder the make upgrade job so that sourc txt files are created before
they are referenced by later pip-tools calls.

local.txt was not generated and doesn't seem to be used by anything,
remove it.