docs: Provied a simple long_description for now.
For some reason the PyPI deploy was inferring a long_description from
somewhere and it had syntax errors so just be explicit so that we can
deploy the rest of the changes.

This is not making things worse because currently there is no
description on PyPI so we're not removing any info.